feat: folder structure changes inside auth folder
