+++ src/js/pages/login/login.js
@@ -3,8 +3,8 @@ import "../../pages/scss/login.scss";
 
 // Import all of Bootstrap’s JS
 import * as bootstrap from "bootstrap";
-import { showToast } from "../showToast.js";
-import authApi from "./AuthApi.js";
+import { showToast } from "../../showToast.js";
+import authApi from "../api/AuthApi.js";
 const authAPI = new authApi();
 
 const loginForm = document.querySelector('form');
