+++ src/js/pages/signup/signUp.js
@@ -3,9 +3,9 @@
 
 // Import all of Bootstrap's JS
 import * as bootstrap from "bootstrap";
-import { showToast } from "../showToast.js";
+import { showToast } from "../../showToast.js";
 
-import authApi from "./AuthApi.js";
+import authApi from "../api/AuthApi.js";
 const authAPI = new authApi();
 
 const signupForm = document.getElementById('signup-form');
