+++ src/pages/forgotOtpVerify.html
@@ -158,6 +158,9 @@
       style="z-index: 1080"
     ></div>
 
-    <script type="module" src="../js/pages/forgotOtpVerify.js"></script>
+    <script
+      type="module"
+      src="../js/pages/forgotPassword/forgotOtpVerify.js"
+    ></script>
   </body>
 </html>
