+++ src/pages/forgotPasswordPage.html
@@ -106,6 +106,9 @@
       style="z-index: 1080"
     ></div>
 
-    <script type="module" src="../js/pages/forgotPassword.js"></script>
+    <script
+      type="module"
+      src="../js/pages/forgotPassword/forgotPassword.js"
+    ></script>
   </body>
 </html>
