+++ src/pages/forgotPasswordReset.html
@@ -75,6 +75,9 @@
       style="z-index: 1080"
     ></div>
 
-    <script type="module" src="../js/pages/forgotPasswordReset.js"></script>
+    <script
+      type="module"
+      src="../js/pages/forgotPassword/forgotPasswordReset.js"
+    ></script>
   </body>
 </html>
