+++ src/pages/loginPage.html
@@ -125,6 +125,6 @@
       style="z-index: 1080"
     ></div>
 
-    <script type="module" src="../js/pages/login.js"></script>
+    <script type="module" src="../js/pages/login/login.js"></script>
   </body>
 </html>
