+++ src/pages/otpPage.html
@@ -168,6 +168,6 @@
       style="z-index: 1080"
     ></div>
 
-    <script type="module" src="../js/pages/otp.js"></script>
+    <script type="module" src="../js/pages/otp/otp.js"></script>
   </body>
 </html>
