+++ src/pages/resetPassword.html
@@ -51,6 +51,9 @@
       style="z-index: 1080"
     ></div>
 
-    <script type="module" src="../js/pages/resetPassword.js"></script>
+    <script
+      type="module"
+      src="../js/pages/resetPassword/resetPassword.js"
+    ></script>
   </body>
 </html>
