+++ src/pages/signUpPage.html
@@ -115,6 +115,6 @@
       style="z-index: 1080"
     ></div>
 
-    <script type="module" src="../js/pages/signUp.js"></script>
+    <script type="module" src="../js/pages/signup/signUp.js"></script>
   </body>
 </html>
